[Guided onboarding] Moved the guided onboarding hook to a different folder in the fleet plugin, also fixed tests
