+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/multi_page_layout/components/confirm_incoming_data_with_preview.tsx
@@ -27,9 +27,7 @@ import type { SearchHit } from '@kbn/es-types';
 
 import styled from 'styled-components';
 
-import { useIsGuidedOnboardingActive } from '../../../../../../integrations/sections/epm/screens/detail/hooks';
-
-import { useStartServices } from '../../../../../../../hooks';
+import { useStartServices, useIsGuidedOnboardingActive } from '../../../../../../../hooks';
 
 import type { PackageInfo } from '../../../../../../../../common';
