+++ x-pack/plugins/fleet/public/applications/integrations/sections/epm/screens/detail/hooks/index.ts
@@ -6,4 +6,3 @@
  */
 
 export { useIsFirstTimeAgentUser } from './use_is_first_time_agent_user';
-export { useIsGuidedOnboardingActive } from './use_is_guided_onboarding_active';
