+++ x-pack/plugins/fleet/public/hooks/index.ts
@@ -28,3 +28,4 @@ export * from './use_agent_policy_refresh';
 export * from './use_package_installations';
 export * from './use_agent_enrollment_flyout_data';
 export * from './use_flyout_context';
+export { useIsGuidedOnboardingActive } from './use_is_guided_onboarding_active';
